fix: resolve golangci-lint findings in postgres adapter
- Check/ignore Close errors explicitly (errcheck)
- Use errors.Is for migrate.ErrNoChange (errorlint)
- Group local imports separately (goimports)
- Make PoolConfig.MaxConns int32 to avoid int64->int32 conversion (gosec G115)

+++ internal/adapters/postgres/pool.go
@@ -16,7 +16,7 @@ import (
 // docs/plano-execucao-wallet-go.md) exists to catch.
 type PoolConfig struct {
 	DSN             string
-	MaxConns        int64
+	MaxConns        int32
 	MaxConnLifetime time.Duration
 	AcquireTimeout  time.Duration
 }
@@ -39,7 +39,7 @@ func NewPool(ctx context.Context, cfg PoolConfig) (*Pool, error) {
 	}
 
 	if cfg.MaxConns > 0 {
-		pgxCfg.MaxConns = int32(cfg.MaxConns)
+		pgxCfg.MaxConns = cfg.MaxConns
 	}
 	if cfg.MaxConnLifetime > 0 {
 		pgxCfg.MaxConnLifetime = cfg.MaxConnLifetime
